refactor(ui): cleanup legacy layout components and organize keyboard elements
