+++ src/app/[locale]/periodicTable/page.tsx
@@ -1,6 +1,6 @@
 import PeriodicTableCards from "@/features/periodic-table/components/PeriodicTableCards";
 import { SubtitleProvider } from "@/shared/contexts/SubtitleContext";
-import Page from "@/shared/components/Page";
+import Page from "@/shared/components/layout/Page";
 import { useTranslations } from "next-intl";
 
 export default function PeriodicTablePage() {
