+++ src/shared/components/brand/Logo.tsx
