+++ src/shared/components/buttons/CloseAllButton.tsx
