+++ src/shared/components/layout/Page.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import React from "react";
-import Header from "@/shared/components/Header";
-import Content from "@/shared/components/Content";
-import Footer from "@/shared/components/Footer";
-import SideArea from "@/shared/components/SideArea";
-import Menu from "@/shared/components/menu/Menu";
+import Header from "./Header";
+import Content from "./Content";
+import Footer from "./Footer";
+import SideArea from "./SideArea";
+import Menu from "../menu/Menu";
 import { useSectionTitle } from "@/shared/hooks/useSectionTitle";
 import { useCollapsedMenu } from "@/shared/hooks/useCollapsedMenu";
 
@@ -13,9 +13,12 @@ export interface PageProps {
   title: string;
   subtitle?: string;
   children?: React.ReactNode;
+  calculatorType?: "standard" | "scientific";
+  className?: string;
+  [key: string]: string | boolean | number | React.ReactNode | undefined;
 }
 
-export default function Page(props: PageProps) {
+export default function Page({ title, children, ...rest }: PageProps) {
   const year = new Date().getFullYear();
   const sectionTitle = useSectionTitle();
   const { collapsed, toggleCollapsed } = useCollapsedMenu();
@@ -33,7 +36,11 @@ export default function Page(props: PageProps) {
       </SideArea>
       <div className="flex flex-col flex-1">
         <Header title={sectionTitle} className="h-16 bg-zinc-100" />
-        <Content title={props.title}>{props.children}</Content>
+        <Content title={title}>
+          {React.isValidElement(children)
+            ? React.cloneElement(children, rest)
+            : children}
+        </Content>
         <Footer year={year} />
       </div>
     </div>
