+++ src/shared/components/menu/constants.ts
@@ -26,7 +26,7 @@ export const MENU_WIDTHS = {
 
 export const MENU_CLASSES = {
   BASE: "sidebar relative flex flex-col h-full bg-zinc-100 dark:bg-neutral-900",
-  ITEM: "flex items-center gap-2 px-4 py-2 rounded-md hover:bg-zinc-200 dark:hover:bg-zinc-800",
+  ITEM: "flex items-center gap-2 px-2 py-2 rounded-md hover:bg-zinc-200 dark:hover:bg-zinc-800",
   ACTIVE_ITEM: "bg-zinc-100 dark:bg-zinc-900",
   SUBMENU: "rounded-xl border border-zinc-200 bg-white shadow-lg dark:bg-neutral-800 dark:border-neutral-700 min-w-[210px]",
   SUBMENU_ITEM: "flex items-center gap-2 w-full px-3 py-2 rounded-md text-2m hover:bg-blue-100 dark:hover:bg-blue-900/40 mt-1",
